Update emmett-2018-rewriting-story-human.md
Added author responses to the editorial critiques of this article, and adjusted the formatting to more clearly denote the additions.  

In the spirit of open & collaboratory publishing, I would request that this version either replace the current version, or be linked to prominently as an additional article in your series.